requirements.txt
streamlit>=1.28.0
ultralytics>=8.0.0
pandas>=1.5.0
numpy>=1.23.0
opencv-python-headless>=4.7.0.72
Pillow>=9.5.0
matplotlib>=3.7.0
seaborn>=0.12.0
torch>=1.13.0
